back-end/requirements.txt
alembic==1.3.0
attrs==19.3.0
Automat==0.8.0
Babel==2.7.0
bleach==3.1.0
blinker==1.4
cached-property==1.5.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
config==0.4.2
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
dominate==2.4.0
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-PageDown==0.2.2
flask-peewee==3.0.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-SSLify==0.1.5
flask-whooshee==0.7.0
Flask-WTF==0.14.2
get==2019.4.13
gevent==1.4.0
greenlet==0.4.15
httpie==1.0.3
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
mysqlclient==1.4.5
parsel==1.5.2
peewee==3.11.2
peewee-migrate==1.1.6
pony==0.7.11
post==2019.4.13
Protego==0.1.15
psycopg2==2.8.4
public==2019.4.13
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.4.2
PyHamcrest==1.9.0
PyMySQL==0.9.3
pyOpenSSL==19.0.0
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
query-string==2019.4.13
queuelib==1.5.0
request==2019.4.13
requests==2.22.0
Scrapy==1.8.0
service-identity==18.1.0
six==1.13.0
SQLAlchemy==1.3.10
Twisted==19.7.0
urllib3==1.25.6
visitor==0.1.3
w3lib==1.21.0
webencodings==0.5.1
Werkzeug==0.16.0
Whoosh==2.7.4
wtf-peewee==3.0.0
WTForms==2.2.1
zope.interface==4.6.0
